FIX: Be careful about overflow with infinite degree
We want the degrees of physical registers to be high enough to not get
insignificant through simplifications, because that would mean they
would be pushed to the stack, so we previously set their degree to be
UINT32_MAX.

But, with coalescing we can add edges to physical registers after
combining them with some virtual register. This would mean that the
degree could overflow.

Checking for overflow and saturating the degree could be a solution, but
here we simply set the initial degree of physical registers to be the
number of virtual registers plus the number of available physical
registers. This means that even if all vregs decrement the degree of a
physical register, it will still be significant.