Have instruction linked list per basic block
Previously I changed it to be per function, this proved to be difficult,
since we need to iterate over block instructions _backwards_ (for
liveness). There was a hack in place, which kept a pointer to the last
instruction, but that didn't allow any easy changes to the basic block
(the last pointer could get invalidated by changes in the instruction
stream).