FIX: Allow peepholes based on RBP
RBP has exactly two definitions in every function:

        ; entry
        mov rbp, rsp
        ...
        pop rbp

But in fact, we know that except for the prologue, it has always the
same value, so we want to optimize based on that.

This would be solved by a more thorough system, which would track
definitions in a better way, but adding a special case helps for now.