FIX: Only enable block merging in the last peephole pass
Well, merging blocks would not exactly keep them basic, which worsened
the result of liveness analysis and, in consequence, register
allocation. Also it was wrong anyways, since we didn't update any
successor/predecessor information or nothing like that.

We can still merge the blocks, but only in the last pass, where we don't
follow up with anything that might depend on the blocks being "basic" or
on correct control flow information.