FIX: Make phi use remove optional
If the phi is trivial, but doesn't use itself, e.g.:

    v35 = phi v19, v19

Then we can't obviously delete the phi from its uses. Thus we still try
the deletion, but don't assert that it happens.

On the other hand, with other deletions we are certain that the use is
there, so we want to assert the deletion.