FIX: Propage resulting value of simplified phi
When value numbering, we save the phi as the last definition of a
variable in merge points. If however, the phi is trivial, e.g.

     v31 = phi v19, v19

We can remove it and use v19 instead. For that however, we need to
propagate that v19 is in fact the value, that should be saved as
canonical.