Use Sreedhar's SSA deconstruction
Sreedhar (Method 1) is more complicated than the two step copying in
predecessors, because he demands that the copies are merged through the
same register, which is not that nice in our SSA representation.

We do the SSA deconstruction in the SSA representation, because of
cycles. Deconstruction of a phi would require copy in predecessor
blocks, but these could not be either not yet present or just being
translated, so adding a copy to the end wouldn't be possible.