Offset used register bitmap by one
We use a single integer with bits corresponding to each physical
register for representing which registers are occupied and not available
for the register being assigned.

Register zero (0) is used as a special register which signifies for
example that no register has been allocated (i.e. for actual spill).
Previously we offset the bitmap by one to not represent R_NONE
there, but we can avoid it and get better looking code.

Though with that we lose one possible physical register in the bitmap.
I.e. for 32 physical registers, 32 bit integer would no longer be able
to represent the bitmap.