Use inverse of spill metric, account for copies
Copy instructions of spilled nodes get deleted (see rewrite_program),
because just the load or store is necessary, so we can subtract the
amount of moves from the cost.

But now the metric can be zero or even negative if all the uses and
definitions are move instructions. Because of this we will now calculate
inverse of the metric, which divides by the degree and thus has no
problems of dividing by zero or negative number.

Just to reflect the actual costs of the operations, we make uses and
definitions twice as costly as moves.